--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -10,7 +10,7 @@ on:
 
 env:
   vcpkg_robotology_TAG: v0.4.0
-  vcpkg_TAG: 222c35e3bcb8f28cd63fc526e591bb4ef6b99e4f
+  vcpkg_TAG: 76a7e9248fb3c57350b559966dcaa2d52a5e4458
   YCM_TAG: v0.11.1
   YARP_TAG: 964bb26fa4791d83d72882711ea7509306248106
   iDynTree_TAG: v1.1.0