--- a/src/Planners/include/BipedalLocomotion/Planners/TimeVaryingDCMPlanner.h
+++ b/src/Planners/include/BipedalLocomotion/Planners/TimeVaryingDCMPlanner.h
@@ -44,6 +44,18 @@ public:
     /**
      * Initialize the planner.
      * @param handler pointer to the parameter handler.
+     * @note the following parameters are required by the class
+     * |           Parameter Name          |    Type    |                                                                   Description                                                                  | Mandatory |
+     * |:---------------------------------:|:----------:|:----------------------------------------------------------------------------------------------------------------------------------------------:|:---------:|
+     * |          `linear_solver`          |  `string`  | Linear solver used by ipopt, the default value is mumps. Please check https://coin-or.github.io/Ipopt/#PREREQUISITES for the available solvers |     No    |
+     * |      `planner_sampling_time`      |  `double`  |                                                          Sampling time of the planner                                                          |    Yes    |
+     * |      `number_of_foot_corners`     |    `int`   |                                      Number of the corner of the polygon used to describe the foot. E.g. 4                                     |    Yes    |
+     * |         `foot_corner_<i>`         | `Vector3d` |             A 3d vector describing the position of the corner w.r.t. frame associated to the foot. `i = 0:number_of_foot_corners`.             |    Yes    |
+     * |         `omega_dot_weight`        |  `double`  |                                                   Weight associated to the $\f\dot{omega}$\f                                                   |    Yes    |
+     * |       `dcm_tracking_weight`       |  `double`  |                                                      Weight associated to the DCM tracking                                                     |    Yes    |
+     * | `omega_dot_rate_of_change_weight` |  `double`  |                                          Weight associated to the rate of change of $\f\dot{omega}$\f                                          |    Yes    |
+     * |    `vrp_rate_of_change_weight`    |  `double`  |                                               Weight associated to the rate of change of the VRP                                               |    Yes    |
+     * |    `dcm_rate_of_change_weight`    |  `double`  |                                               Weight associated to the rate of change of the DCM                                               |    Yes    |
      * @return true in case of success/false otherwise.
      */
      bool initialize(std::shared_ptr<ParametersHandler::IParametersHandler> handler) override;