--- a/src/Planners/src/TimeVaryingDCMPlanner.cpp
+++ b/src/Planners/src/TimeVaryingDCMPlanner.cpp
@@ -569,6 +569,10 @@ bool TimeVaryingDCMPlanner::initialize(std::shared_ptr<ParametersHandler::IParam
         }
     }
 
+    // get the linear solver used by ipopt. This parameter is optional. The default value is mumps
+    std::string linearSolver;
+    if (handler->getParameter("linear_solver", linearSolver))
+        m_pimpl->optiSettings.ipoptLinearSolver = linearSolver;
 
     bool okCostFunctions = true;
     okCostFunctions &= handler->getParameter("omega_dot_weight", m_pimpl->optiSettings.omegaDotWeight);