@traversaro 2020-10-07
> 
> 
> CI is failing on windows with this error that this not related with this PR
> 
> ```
> 2020-10-03T02:13:45.5270628Z -- Acquiring MSYS2...
> 2020-10-03T02:13:45.5271845Z -- Downloading https://sourceforge.net/projects/msys2/files/Base/x86_64/msys2-base-x86_64-20190524.tar.xz/download...
> 2020-10-03T02:13:45.5273331Z -- Downloading http://repo.msys2.org/msys/x86_64/msys2-keyring-r21.b39fb11-1-any.pkg.tar.xz...
> 2020-10-03T02:13:45.5275261Z -- Downloading http://repo.msys2.org/msys/x86_64/msys2-keyring-r21.b39fb11-1-any.pkg.tar.xz... Failed. Status: 28;"Timeout was reached"
> 2020-10-03T02:13:45.5276616Z CMake Error at scripts/cmake/vcpkg_download_distfile.cmake:176 (message):
> 2020-10-03T02:13:45.5277185Z       
> 2020-10-03T02:13:45.5277545Z       Failed to download file.
> 2020-10-03T02:13:45.5278263Z       If you use a proxy, please set the HTTPS_PROXY and HTTP_PROXY environment
> 2020-10-03T02:13:45.5282223Z       variables to "***your-proxy-ip-address:port/".
> 2020-10-03T02:13:45.5283011Z       Otherwise, please submit an issue at https://github.com/Microsoft/vcpkg/issues
> 2020-10-03T02:13:45.5283441Z 
> 2020-10-03T02:13:45.5283740Z Call Stack (most recent call first):
> 2020-10-03T02:13:45.5284241Z   scripts/cmake/vcpkg_acquire_msys.cmake:86 (vcpkg_download_distfile)
> 2020-10-03T02:13:45.5284848Z   scripts/cmake/vcpkg_fixup_pkgconfig.cmake:281 (vcpkg_acquire_msys)
> 2020-10-03T02:13:45.5285431Z   ports/cppad/portfile.cmake:39 (vcpkg_fixup_pkgconfig)
> 2020-10-03T02:13:45.5285909Z   scripts/ports.cmake:79 (include)
> 2020-10-03T02:13:45.5286181Z 
> 2020-10-03T02:13:45.5286311Z 
> 2020-10-03T02:13:45.5286706Z Error: Building package cppad:x64-windows failed with: BUILD_FAILED
> 2020-10-03T02:13:45.5287284Z Please ensure you're using the latest portfiles with `.\vcpkg update`, then
> 2020-10-03T02:13:45.5287941Z submit an issue at https://github.com/Microsoft/vcpkg/issues including:
> 2020-10-03T02:13:45.5288478Z   Package: cppad:x64-windows
> 2020-10-03T02:13:45.5288819Z   Vcpkg version: 2020.06.15-nohash
> ```
> 
> Associated issue: [microsoft/vcpkg#13286](https://github.com/microsoft/vcpkg/issues/13286)
> cc @traversaro

That error was fixed in https://github.com/microsoft/vcpkg/pull/13889 . I am generating a new archive in https://github.com/robotology/robotology-superbuild-dependencies-vcpkg, but indeed there are regressions. For the time being I suggest to bump the vcpkg tag to  76a7e9248fb3c57350b559966dcaa2d52a5e4458 as done in https://github.com/robotology/robotology-vcpkg-ports/pull/16 .