@GiulioRomualdi 2020-10-07
Thank you @traversaro. In https://github.com/dic-iit/bipedal-locomotion-framework/pull/127/commits/e82a5e32c43de688fddf7b4ae8c1a5a3bb3ca3df I bumped the `vcpkg` commit.  Let's see if it solves the problem.